Moving likelihood functions to likelihood.py
Moved from defining functions inside fitter functions within fitters.py
to defining them in likelihood.py for clarity
