+++ eureka/S5_lightcurve_fitting/likelihood.py
@@ -1,4 +1,45 @@
 import numpy as np
+from scipy.stats import norm
+
+def ln_like(theta, lc, model, pmin, pmax, freenames):
+    # params[ifreepars] = freepars
+    ilow = np.where(theta < pmin)
+    ihi = np.where(theta > pmax)
+    theta[ilow] = pmin[ilow]
+    theta[ihi] = pmax[ihi]
+    model.update(theta, freenames)
+    model_lc = model.eval()
+    residuals = (lc.flux - model_lc) #/ lc.unc
+    ln_like_val = (-0.5 * (np.sum((residuals / lc.unc) ** 2+ np.log(2.0 * np.pi * (lc.unc) ** 2))))
+    if len(ilow[0]) + len(ihi[0]) > 0: ln_like_val = -np.inf
+    return ln_like_val
+
+def lnprior(theta, pmin, pmax):
+    lnprior_prob = 0.
+    n = len(theta)
+    for i in range(n):
+        if np.logical_or(theta[i] < pmin[i],
+                                theta[i] > pmax[i]): lnprior_prob += - np.inf
+    return lnprior_prob
+
+def lnprob(theta, lc, model, pmin, pmax, freenames):
+    ln_like_val = ln_like(theta, lc, model, pmin, pmax, freenames)
+    lp = lnprior(theta, pmin, pmax)
+    return ln_like_val + lp
+
+#PRIOR TRANSFORMATION TODO: ADD GAUSSIAN PRIORS
+def transform_uniform(x, a, b):
+    return a + (b - a) * x
+
+def transform_normal(x, mu, sigma):
+    return norm.ppf(x, loc=mu, scale=sigma)
+
+def ptform(theta, pmin, pmax):
+    p = np.zeros_like(theta)
+    n = len(theta)
+    for i in range(n):
+        p[i] = transform_uniform(theta[i], pmin[i], pmax[i])
+    return p
 
 def computeRedChiSq(lc, model, meta, freenames):
 	model_lc = model.eval()
